Split karva_cli lib.rs into per-command modules
